src/requirements.txt
asgiref==3.2.7
DateTime==4.3
Django==3.0.6
django-cors-headers==3.2.1
django-material==1.6.3
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
zope.interface==4.7.1
